sefs-project/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
watchdog==3.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
PyPDF2==3.0.1
python-multipart==0.0.6
numpy==1.26.2
pandas==2.1.3
crewai==0.28.0
crewai-tools==0.2.6
websockets==12.0
python-dotenv==1.0.0
langchain-google-genai>=2.0.0
langchain-openai>=0.0.5
